+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - 'POST /consent' handler accepts `client_id` and `scope` parameters in request
 body and creates grants with the corresponding values.
-- 'consent' view is passed `application` and `scope` parameters via `res.locals`.
+- 'consent' view is passed `user`, `application`, and `scope` parameters via
+`res.locals`.
 - 'consent' prompt accepts `client` and `scope` options via `res.locals` and
 forwards them as `client_id` and `scope` query parameters to the 'GET /consent'
 endpoint.